Fix: fix circular import
