+++ src/careamics/lightning/lightning_modules/care_module.py
@@ -11,7 +11,7 @@ from torchmetrics import MetricCollection
 from careamics.config import CAREAlgorithm, N2NAlgorithm
 from careamics.config.factories.algorithm_factory import algorithm_factory
 from careamics.config.support import SupportedLoss
-from careamics.dataset.dataset import ImageRegionData
+from careamics.dataset.image_region_data import ImageRegionData
 from careamics.lightning.data_module import TrainValData, TrainValSplitData
 from careamics.lightning.metrics import SIPSNR
 from careamics.losses import mae_loss, mse_loss
