+++ src/careamics/lightning/lightning_modules/n2v_module.py
@@ -10,7 +10,7 @@ from torchmetrics import MetricCollection
 
 from careamics.config import N2VAlgorithm
 from careamics.config.factories.algorithm_factory import algorithm_factory
-from careamics.dataset.dataset import ImageRegionData
+from careamics.dataset.image_region_data import ImageRegionData
 from careamics.lightning.data_module import TrainValData, TrainValSplitData
 from careamics.lightning.metrics import SIPSNR
 from careamics.losses import n2v_loss
