+++ src/careamics/lightning/prediction/stitch_prediction.py
@@ -7,7 +7,7 @@ from typing import Literal
 import numpy as np
 from numpy.typing import NDArray
 
-from careamics.dataset.dataset import ImageRegionData
+from careamics.dataset.image_region_data import ImageRegionData
 from careamics.dataset.patching_strategies import TileSpecs
 from careamics.utils.reshape_array import restore_array
